creator-analysis-main/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
httpx==0.28.1
